refactor(app): correct dir name
