+++ app/lib/common/pages/drug/widgets/annotation_card.dart
