+++ app/lib/common/pages/drug/widgets/module.dart
