+++ app/lib/common/pages/drug/widgets/recommendation_card.dart
