+++ app/lib/search/module.dart
@@ -1,11 +1,11 @@
 import 'package:auto_route/auto_route.dart';
 
-import '../common/pages/drugs/drug.dart';
+import '../common/pages/drug/drug.dart';
 import 'pages/search.dart';
 
-export '../common/pages/drugs/cubit.dart';
+export '../common/pages/drug/cubit.dart';
 // We need to expose all pages for AutoRouter
-export '../common/pages/drugs/drug.dart';
+export '../common/pages/drug/drug.dart';
 export 'pages/cubit.dart';
 export 'pages/search.dart';
